requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# PDF Processing
PyPDF2==3.0.1
python-docx==0.8.11

# NLP and Text Processing
spacy==3.7.2
nltk==3.8.1

# Machine Learning and Similarity Matching
scikit-learn==1.3.2
numpy==1.26.2
scipy==1.11.4

# Data Processing
pandas==2.1.3

# API Requests
requests==2.31.0
httpx==0.25.2

# Environment Variables
python-dotenv==1.0.0

# Database (MySQL)
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.12.0
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1

# Logging and Monitoring
python-multipart==0.0.6
